// src/api/garage.d.ts
export interface components {
    schemas: {
// ...
        BlockError: {
            blockHash: string;
            /** Format: int64 */
            errorCount: number;
            /** Format: int64 */
            lastTrySecsAgo: number;
            /** Format: int64 */
            nextTryInSecs: number;
            /** Format: int64 */
            refcount: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}